+++ thanksbucket/src/main/java/com/thanksbucket/security/authentication/www/session/SessionAuthenticationFailureHandler.java
@@ -1,4 +1,4 @@
-package com.thanksbucket.security.authentication.www;
+package com.thanksbucket.security.authentication.www.session;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
@@ -11,11 +11,9 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.web.authentication.AuthenticationFailureHandler;
-import org.springframework.stereotype.Component;
 
 import java.io.IOException;
 
-@Component
 @Slf4j
 public class SessionAuthenticationFailureHandler implements AuthenticationFailureHandler {
