+++ thanksbucket/src/main/java/com/thanksbucket/security/authentication/www/session/SessionAuthenticationSuccessHandler.java
@@ -1,4 +1,4 @@
-package com.thanksbucket.security.authentication.www;
+package com.thanksbucket.security.authentication.www.session;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
@@ -16,11 +16,9 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
 import org.springframework.security.web.context.HttpSessionSecurityContextRepository;
-import org.springframework.stereotype.Component;
 
 import java.io.IOException;
 
-@Component
 @Slf4j
 public class SessionAuthenticationSuccessHandler implements AuthenticationSuccessHandler {
     private final ObjectMapper objectMapper = new ObjectMapper();
